models/requirements.txt
langchain
langchain-google-genai
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
openai
huggingface-hub
transformers
python-dotenv
numpy
scikit-learn
